add-database

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -1,5 +1,6 @@
 package ru.yandex.practicum.filmorate.service;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
@@ -13,10 +14,10 @@ import ru.yandex.practicum.filmorate.model.Genre;
 import ru.yandex.practicum.filmorate.model.Mpa;
 
 import java.sql.SQLException;
-import java.util.Collection;
 import java.util.List;
 
 @Service
+@Slf4j
 public class FilmService {
     private final FilmDbStorage filmStorage;
     private final MpaDbStorage mpaStorage;
@@ -25,7 +26,7 @@ public class FilmService {
 
     @Autowired
     public FilmService(FilmDbStorage filmStorage, MpaDbStorage mpaStorage,
-                       GenresStorage genresStorage, UserDbStorage userStorage ) {
+                       GenresStorage genresStorage, UserDbStorage userStorage) {
         this.filmStorage = filmStorage;
         this.mpaStorage = mpaStorage;
         this.genresStorage = genresStorage;
@@ -34,17 +35,17 @@ public class FilmService {
 
     public List<Film> getAllFilms() {
         List<Film> films = filmStorage.getFilmMap();
-        for(Film film:films){
+        for (Film film : films) {
             film.setGenres(genresStorage.getGenresByFilmId(film.getId()));
         }
         return films;
     }
 
-    public Film get(int filmId)  {
+    public Film get(int filmId) {
         Film film;
         try {
             film = filmStorage.get(filmId);
-        } catch (EmptyResultDataAccessException e){
+        } catch (EmptyResultDataAccessException e) {
             throw new NotFoundException("Фильм не найден");
         }
         int mpaId = film.getMpa().getId();
@@ -68,15 +69,21 @@ public class FilmService {
     }
 
     public Film updateFilm(Film film) {
-        filmStorage.updateFilm(film);
-        genresStorage.deleteGenresForFilm(film.getId());
-        genresStorage.setGenresForFilm(film.getId());
-        filmStorage.updateFilm(film);
+        if (!(getAllFilms().isEmpty())) {
+            if (!getAllFilms().contains(film)) {
+                log.info("Фильма не существует");
+                throw new NotFoundException("Фильм с id " + film.getId() + "не найден");
+            }
+            filmStorage.updateFilm(film);
+            genresStorage.deleteGenresForFilm(film.getId());
+            genresStorage.setGenresForFilm(film.getId());
+            filmStorage.updateFilm(film);
+        }
         return film;
     }
 
     public void addLike(int userId, int filmId) throws SQLException {
-            filmStorage.addLike(userId, filmId);
+        filmStorage.addLike(userId, filmId);
     }
 
     public void deleteLike(int userId, int filmId) throws SQLException {
@@ -85,7 +92,7 @@ public class FilmService {
 
     public List<Film> raitingFilm(int count) {
         List<Film> films = filmStorage.raitingFilm(count);
-        for(Film film:films){
+        for (Film film : films) {
             film.setGenres(genresStorage.getGenresByFilmId(film.getId()));
         }
         return films;
